Get the cell assignment probabilities of a cellassign fit

Get the MLE cell type assignment probabilities for each cell

cellprobs(x)

# S3 method for cellassign
cellprobs(x)

Arguments

x
An object of class cellassign returned by a call to cellassign(...)

Value

A cell by celltype matrix with assignment probabilities

Examples

data(example_cellassign_fit)
cellprobs(example_cellassign_fit)

#>               Group1       Group2
#>   [1,]  1.000000e+00 5.631750e-42
#>   [2,] 3.143573e-161 1.000000e+00
#>   [3,]  2.452796e-48 1.000000e+00
#>   [4,]  1.000000e+00 2.343295e-29
#>   [5,]  1.000000e+00 1.563717e-33
#>   [6,]  1.000000e+00 2.310088e-12
#>   [7,]  1.000000e+00 9.540475e-62
#>   [8,]  2.678417e-70 1.000000e+00
#>   [9,]  1.000000e+00 9.433894e-26
#>  [10,]  1.000000e+00 2.220594e-51
#>  [11,]  3.349618e-78 1.000000e+00
#>  [12,]  1.000000e+00 3.359494e-17
#>  [13,]  8.484366e-44 1.000000e+00
#>  [14,]  1.000000e+00 1.471530e-19
#>  [15,]  1.000000e+00 4.401413e-33
#>  [16,]  1.000000e+00 2.519241e-10
#>  [17,]  2.152278e-71 1.000000e+00
#>  [18,]  1.000000e+00 1.431516e-26
#>  [19,]  8.485051e-54 1.000000e+00
#>  [20,]  1.553482e-76 1.000000e+00
#>  [21,]  6.887468e-55 1.000000e+00
#>  [22,]  1.000000e+00 3.165604e-28
#>  [23,]  1.021700e-19 1.000000e+00
#>  [24,]  1.000000e+00 1.622092e-32
#>  [25,]  1.000000e+00 1.108399e-30
#>  [26,]  1.745562e-45 1.000000e+00
#>  [27,]  1.000000e+00 1.759241e-61
#>  [28,]  1.939442e-77 1.000000e+00
#>  [29,]  1.000000e+00 3.487263e-36
#>  [30,]  1.187917e-55 1.000000e+00
#>  [31,]  1.749304e-32 1.000000e+00
#>  [32,]  4.518632e-81 1.000000e+00
#>  [33,]  1.000000e+00 2.431815e-18
#>  [34,]  1.000000e+00 2.697196e-27
#>  [35,]  2.101276e-45 1.000000e+00
#>  [36,]  8.693355e-32 1.000000e+00
#>  [37,]  1.000000e+00 8.129691e-31
#>  [38,]  4.729411e-43 1.000000e+00
#>  [39,]  1.309477e-63 1.000000e+00
#>  [40,]  2.463022e-93 1.000000e+00
#>  [41,]  1.000000e+00 3.456708e-19
#>  [42,]  3.714639e-98 1.000000e+00
#>  [43,]  3.351410e-86 1.000000e+00
#>  [44,] 6.389950e-121 1.000000e+00
#>  [45,]  1.629260e-69 1.000000e+00
#>  [46,]  1.000000e+00 9.653211e-17
#>  [47,]  6.525926e-87 1.000000e+00
#>  [48,]  1.000000e+00 1.753990e-16
#>  [49,]  8.061459e-44 1.000000e+00
#>  [50,]  1.786365e-90 1.000000e+00
#>  [51,]  1.000000e+00 4.471191e-18
#>  [52,]  1.000000e+00 1.634998e-44
#>  [53,]  4.065618e-76 1.000000e+00
#>  [54,]  1.000000e+00 1.486390e-33
#>  [55,]  1.000000e+00 4.735952e-31
#>  [56,]  1.000000e+00 5.529861e-37
#>  [57,]  1.616758e-85 1.000000e+00
#>  [58,]  4.205798e-55 1.000000e+00
#>  [59,]  1.000000e+00 3.501183e-42
#>  [60,]  1.000000e+00 2.220353e-21
#>  [61,]  2.717309e-88 1.000000e+00
#>  [62,]  1.000000e+00 1.007551e-24
#>  [63,]  5.646554e-40 1.000000e+00
#>  [64,]  2.999852e-40 1.000000e+00
#>  [65,]  1.000000e+00 7.006973e-22
#>  [66,]  1.334329e-68 1.000000e+00
#>  [67,]  1.000000e+00 1.169466e-30
#>  [68,]  1.000000e+00 4.525840e-33
#>  [69,] 3.096411e-104 1.000000e+00
#>  [70,]  1.000000e+00 2.850298e-35
#>  [71,]  1.000000e+00 1.710052e-21
#>  [72,]  1.000000e+00 1.397341e-37
#>  [73,] 1.260048e-142 1.000000e+00
#>  [74,]  7.117283e-85 1.000000e+00
#>  [75,]  1.503771e-31 1.000000e+00
#>  [76,]  1.000000e+00 3.511943e-53
#>  [77,]  1.000000e+00 3.328239e-28
#>  [78,]  1.000000e+00 7.340310e-36
#>  [79,] 4.717143e-122 1.000000e+00
#>  [80,]  1.000000e+00 3.575386e-22
#>  [81,]  6.955760e-91 1.000000e+00
#>  [82,]  1.229655e-27 1.000000e+00
#>  [83,]  3.948177e-33 1.000000e+00
#>  [84,]  3.048179e-56 1.000000e+00
#>  [85,]  1.000000e+00 4.659612e-69
#>  [86,]  9.240999e-58 1.000000e+00
#>  [87,]  5.677020e-66 1.000000e+00
#>  [88,]  1.000000e+00 1.273649e-79
#>  [89,]  1.582613e-54 1.000000e+00
#>  [90,]  6.033033e-57 1.000000e+00
#>  [91,] 3.200480e-118 1.000000e+00
#>  [92,]  6.285678e-33 1.000000e+00
#>  [93,]  1.000000e+00 1.146502e-20
#>  [94,]  1.000000e+00 1.137067e-42
#>  [95,]  2.950847e-43 1.000000e+00
#>  [96,]  1.420613e-48 1.000000e+00
#>  [97,]  1.000000e+00 1.387482e-43
#>  [98,]  1.168605e-37 1.000000e+00
#>  [99,]  1.000000e+00 1.646072e-25
#> [100,]  5.790859e-47 1.000000e+00
#> [101,]  1.000000e+00 4.384632e-47
#> [102,] 5.296492e-118 1.000000e+00
#> [103,] 6.641791e-107 1.000000e+00
#> [104,]  1.274184e-65 1.000000e+00
#> [105,]  2.240147e-55 1.000000e+00
#> [106,]  1.000000e+00 2.512884e-53
#> [107,]  1.000000e+00 4.854778e-34
#> [108,]  1.034917e-69 1.000000e+00
#> [109,]  1.000000e+00 8.667301e-65
#> [110,]  1.000000e+00 1.399332e-09
#> [111,]  1.000000e+00 1.235603e-18
#> [112,]  1.000000e+00 3.792226e-12
#> [113,]  1.721046e-25 1.000000e+00
#> [114,]  5.933363e-17 1.000000e+00
#> [115,]  4.830675e-35 1.000000e+00
#> [116,]  1.000000e+00 7.550915e-26
#> [117,]  1.000000e+00 4.097481e-27
#> [118,] 1.488814e-136 1.000000e+00
#> [119,]  2.190167e-91 1.000000e+00
#> [120,]  1.000000e+00 6.179423e-35
#> [121,]  1.320571e-24 1.000000e+00
#> [122,]  2.030756e-39 1.000000e+00
#> [123,]  1.000000e+00 2.264624e-61
#> [124,]  1.000000e+00 1.150427e-19
#> [125,]  1.000000e+00 4.423178e-22
#> [126,]  3.884162e-64 1.000000e+00
#> [127,]  1.000000e+00 3.378723e-31
#> [128,]  1.000000e+00 5.343542e-26
#> [129,]  8.123727e-45 1.000000e+00
#> [130,]  1.018962e-98 1.000000e+00
#> [131,]  4.635609e-30 1.000000e+00
#> [132,]  2.845628e-23 1.000000e+00
#> [133,]  1.000000e+00 5.121548e-37
#> [134,]  4.846261e-31 1.000000e+00
#> [135,]  1.000000e+00 1.120397e-17
#> [136,]  1.000000e+00 1.038913e-24
#> [137,]  1.000000e+00 4.763773e-59
#> [138,]  1.064475e-35 1.000000e+00
#> [139,]  9.999999e-01 6.621897e-08
#> [140,]  1.000000e+00 5.737150e-35
#> [141,]  1.000000e+00 1.960484e-50
#> [142,]  1.000000e+00 3.213031e-15
#> [143,]  2.970954e-58 1.000000e+00
#> [144,]  3.591204e-47 1.000000e+00
#> [145,]  1.000000e+00 1.428657e-51
#> [146,]  3.087493e-49 1.000000e+00
#> [147,]  1.000000e+00 1.601600e-50
#> [148,]  1.000000e+00 1.121271e-52
#> [149,]  4.777437e-52 1.000000e+00
#> [150,]  1.000000e+00 7.543704e-20
#> [151,]  1.000000e+00 6.066853e-18
#> [152,]  1.000000e+00 1.736146e-35
#> [153,]  1.000000e+00 1.370096e-40
#> [154,]  1.221422e-36 1.000000e+00
#> [155,]  1.593513e-95 1.000000e+00
#> [156,]  1.314730e-60 1.000000e+00
#> [157,]  1.000000e+00 6.850922e-23
#> [158,]  1.468434e-25 1.000000e+00
#> [159,]  8.736578e-12 1.000000e+00
#> [160,]  1.000000e+00 3.614686e-51
#> [161,]  2.591087e-45 1.000000e+00
#> [162,]  1.000000e+00 2.206213e-51
#> [163,]  1.000000e+00 1.647129e-17
#> [164,]  4.809920e-08 1.000000e+00
#> [165,]  3.835634e-85 1.000000e+00
#> [166,]  1.000000e+00 3.342410e-38
#> [167,]  1.000000e+00 8.667762e-47
#> [168,]  1.000000e+00 2.853815e-41
#> [169,]  1.000000e+00 2.765288e-49
#> [170,]  1.771338e-20 1.000000e+00
#> [171,]  1.138978e-37 1.000000e+00
#> [172,]  1.000000e+00 4.722906e-39
#> [173,]  1.000000e+00 8.926998e-42
#> [174,]  1.000000e+00 4.155089e-37
#> [175,]  1.068036e-34 1.000000e+00
#> [176,]  6.078877e-93 1.000000e+00
#> [177,]  1.000000e+00 6.751954e-33
#> [178,]  1.000000e+00 1.343224e-15
#> [179,]  1.529101e-35 1.000000e+00
#> [180,]  6.548048e-24 1.000000e+00
#> [181,]  1.159574e-16 1.000000e+00
#> [182,]  2.219527e-63 1.000000e+00
#> [183,]  5.835370e-59 1.000000e+00
#> [184,]  9.950811e-24 1.000000e+00
#> [185,]  3.745371e-69 1.000000e+00
#> [186,]  1.000000e+00 2.144458e-43
#> [187,]  1.000000e+00 1.475701e-65
#> [188,]  2.132603e-35 1.000000e+00
#> [189,]  4.921920e-28 1.000000e+00
#> [190,] 1.434230e-110 1.000000e+00
#> [191,]  1.863013e-86 1.000000e+00
#> [192,]  1.000000e+00 1.461720e-36
#> [193,]  1.000000e+00 3.872142e-56
#> [194,]  1.000000e+00 1.892276e-23
#> [195,]  1.000000e+00 9.410072e-27
#> [196,]  1.000000e+00 1.085373e-53
#> [197,] 4.761595e-145 1.000000e+00
#> [198,]  1.000000e+00 4.700306e-68
#> [199,]  1.000000e+00 6.445129e-15
#> [200,]  4.059595e-40 1.000000e+00
#> [201,]  4.505210e-45 1.000000e+00
#> [202,]  1.324050e-56 1.000000e+00
#> [203,]  8.578708e-53 1.000000e+00
#> [204,]  1.000000e+00 3.308755e-15
#> [205,]  2.775275e-48 1.000000e+00
#> [206,]  8.480883e-94 1.000000e+00
#> [207,]  4.776222e-56 1.000000e+00
#> [208,]  3.809244e-54 1.000000e+00
#> [209,]  1.000000e+00 1.169642e-12
#> [210,]  1.000000e+00 1.217238e-27
#> [211,]  1.000000e+00 8.913462e-10
#> [212,]  4.590987e-67 1.000000e+00
#> [213,]  1.000000e+00 7.515732e-60
#> [214,]  1.000000e+00 6.905329e-26
#> [215,]  1.145111e-91 1.000000e+00
#> [216,]  1.000000e+00 7.969781e-34
#> [217,] 2.776762e-108 1.000000e+00
#> [218,]  1.000000e+00 8.700499e-12
#> [219,]  1.137934e-28 1.000000e+00
#> [220,]  9.878770e-59 1.000000e+00
#> [221,]  1.000000e+00 1.454543e-16
#> [222,]  7.027761e-73 1.000000e+00
#> [223,]  3.687015e-72 1.000000e+00
#> [224,]  1.000000e+00 1.289574e-33
#> [225,]  1.881130e-23 1.000000e+00
#> [226,] 3.607736e-100 1.000000e+00
#> [227,]  1.000000e+00 3.564174e-13
#> [228,]  1.660289e-85 1.000000e+00
#> [229,]  1.000000e+00 1.454038e-14
#> [230,] 1.622793e-115 1.000000e+00
#> [231,]  1.000000e+00 2.982337e-31
#> [232,]  1.000000e+00 1.003590e-12
#> [233,] 6.753146e-135 1.000000e+00
#> [234,]  1.000000e+00 2.524630e-42
#> [235,]  1.000000e+00 3.287118e-16
#> [236,]  4.126660e-29 1.000000e+00
#> [237,]  1.943415e-50 1.000000e+00
#> [238,]  2.026917e-60 1.000000e+00
#> [239,]  1.000000e+00 7.237972e-17
#> [240,] 1.647078e-131 1.000000e+00
#> [241,]  3.341681e-26 1.000000e+00
#> [242,] 8.716463e-250 1.000000e+00
#> [243,]  1.000000e+00 1.445658e-38
#> [244,]  1.000000e+00 1.589918e-25
#> [245,]  1.000000e+00 5.656980e-11
#> [246,]  4.202757e-49 1.000000e+00
#> [247,]  1.000000e+00 6.844958e-16
#> [248,] 1.298101e-100 1.000000e+00
#> [249,]  1.000000e+00 1.631915e-63
#> [250,]  3.430747e-22 1.000000e+00
#> [251,]  1.000000e+00 5.399287e-22
#> [252,]  1.000000e+00 3.554406e-39
#> [253,]  8.091648e-62 1.000000e+00
#> [254,]  1.000000e+00 2.596309e-47
#> [255,]  3.662316e-38 1.000000e+00
#> [256,]  2.675018e-57 1.000000e+00
#> [257,]  1.000000e+00 3.586525e-25
#> [258,]  1.000000e+00 6.002375e-21
#> [259,] 1.253378e-106 1.000000e+00
#> [260,]  1.000000e+00 3.062477e-55
#> [261,]  5.112009e-37 1.000000e+00
#> [262,]  1.578828e-71 1.000000e+00
#> [263,]  1.053715e-48 1.000000e+00
#> [264,]  6.172314e-26 1.000000e+00
#> [265,]  2.695835e-22 1.000000e+00
#> [266,]  1.000000e+00 7.930118e-25
#> [267,]  1.627809e-44 1.000000e+00
#> [268,]  1.000000e+00 2.787612e-23
#> [269,]  3.697720e-91 1.000000e+00
#> [270,]  1.000000e+00 1.767548e-39
#> [271,]  1.000000e+00 6.315835e-34
#> [272,]  1.000000e+00 2.080406e-36
#> [273,]  3.642171e-50 1.000000e+00
#> [274,]  1.000000e+00 5.534272e-29
#> [275,]  1.000000e+00 3.178372e-33
#> [276,]  1.000000e+00 1.174485e-29
#> [277,]  7.101860e-60 1.000000e+00
#> [278,]  1.000000e+00 7.970267e-44
#> [279,]  1.000000e+00 1.656121e-28
#> [280,]  2.775973e-94 1.000000e+00
#> [281,]  1.000000e+00 3.342194e-10
#> [282,]  1.000000e+00 1.576725e-45
#> [283,]  3.535600e-56 1.000000e+00
#> [284,]  1.199988e-83 1.000000e+00
#> [285,]  1.000000e+00 1.016124e-34
#> [286,]  3.000652e-75 1.000000e+00
#> [287,]  3.069948e-40 1.000000e+00
#> [288,] 8.085603e-168 1.000000e+00
#> [289,]  1.000000e+00 1.226053e-43
#> [290,]  1.000000e+00 9.902991e-47
#> [291,]  6.949419e-61 1.000000e+00
#> [292,]  1.000000e+00 4.086198e-28
#> [293,]  5.460616e-72 1.000000e+00
#> [294,]  1.000000e+00 3.152457e-16
#> [295,]  1.000000e+00 9.300054e-25
#> [296,]  1.000000e+00 5.295175e-42
#> [297,] 2.536973e-140 1.000000e+00
#> [298,]  1.000000e+00 6.486411e-32
#> [299,] 8.463362e-109 1.000000e+00
#> [300,]  1.000000e+00 3.219465e-58
#> [301,]  2.949200e-57 1.000000e+00
#> [302,]  7.861341e-54 1.000000e+00
#> [303,]  1.000000e+00 4.031322e-51
#> [304,]  1.000000e+00 1.395555e-41
#> [305,]  1.000000e+00 6.351407e-32
#> [306,] 5.368446e-129 1.000000e+00
#> [307,] 3.564738e-141 1.000000e+00
#> [308,]  6.255883e-40 1.000000e+00
#> [309,]  6.711779e-58 1.000000e+00
#> [310,]  1.000000e+00 4.019208e-24
#> [311,]  1.000000e+00 6.089636e-13
#> [312,]  2.926214e-50 1.000000e+00
#> [313,]  1.339702e-42 1.000000e+00
#> [314,]  1.000000e+00 1.890072e-10
#> [315,]  1.111588e-85 1.000000e+00
#> [316,]  4.731331e-19 1.000000e+00
#> [317,]  3.718238e-49 1.000000e+00
#> [318,]  3.697509e-35 1.000000e+00
#> [319,]  6.914359e-73 1.000000e+00
#> [320,]  7.379184e-36 1.000000e+00
#> [321,]  1.082378e-37 1.000000e+00
#> [322,]  5.289100e-50 1.000000e+00
#> [323,]  1.781855e-50 1.000000e+00
#> [324,]  1.000000e+00 7.612976e-25
#> [325,]  6.366032e-66 1.000000e+00
#> [326,]  2.543921e-21 1.000000e+00
#> [327,]  1.000000e+00 3.587664e-22
#> [328,]  2.645993e-80 1.000000e+00
#> [329,]  8.119849e-79 1.000000e+00
#> [330,] 9.611796e-134 1.000000e+00
#> [331,]  9.037519e-55 1.000000e+00
#> [332,]  5.636254e-60 1.000000e+00
#> [333,]  1.000000e+00 6.662567e-45
#> [334,]  6.097021e-65 1.000000e+00
#> [335,]  1.000000e+00 3.134135e-18
#> [336,]  1.000000e+00 6.428768e-30
#> [337,]  1.795134e-90 1.000000e+00
#> [338,]  5.716078e-17 1.000000e+00
#> [339,] 3.446149e-211 1.000000e+00
#> [340,]  3.039197e-61 1.000000e+00
#> [341,]  1.000000e+00 1.587719e-35
#> [342,]  3.506908e-18 1.000000e+00
#> [343,]  1.463915e-65 1.000000e+00
#> [344,]  1.000000e+00 3.663819e-31
#> [345,]  1.000000e+00 7.608836e-46
#> [346,]  1.538677e-38 1.000000e+00
#> [347,]  1.000000e+00 7.815718e-27
#> [348,]  1.000000e+00 1.308859e-24
#> [349,]  1.000000e+00 7.086025e-14
#> [350,]  2.220187e-10 1.000000e+00
#> [351,]  1.000000e+00 2.950612e-43
#> [352,]  1.000000e+00 2.906501e-29
#> [353,]  9.999997e-01 3.005187e-07
#> [354,]  1.000000e+00 5.698516e-10
#> [355,]  1.000000e+00 4.511061e-45
#> [356,]  1.000000e+00 2.723883e-37
#> [357,]  1.509010e-31 1.000000e+00
#> [358,]  6.876074e-51 1.000000e+00
#> [359,]  6.567678e-50 1.000000e+00
#> [360,]  1.000000e+00 3.532460e-22
#> [361,] 1.844008e-143 1.000000e+00
#> [362,]  1.000000e+00 2.191923e-28
#> [363,] 8.878094e-188 1.000000e+00
#> [364,]  9.345463e-69 1.000000e+00
#> [365,]  1.000000e+00 1.006224e-30
#> [366,]  3.788105e-53 1.000000e+00
#> [367,]  1.000000e+00 1.481298e-17
#> [368,]  1.000000e+00 9.158644e-74
#> [369,]  2.058294e-55 1.000000e+00
#> [370,]  1.000000e+00 1.292954e-41
#> [371,]  4.316576e-46 1.000000e+00
#> [372,]  1.000000e+00 2.506043e-16
#> [373,]  1.000000e+00 1.035943e-09
#> [374,]  1.000000e+00 8.257822e-56
#> [375,]  1.000000e+00 8.907891e-42
#> [376,]  1.000000e+00 5.963963e-27
#> [377,] 1.739199e-116 1.000000e+00
#> [378,]  6.129768e-68 1.000000e+00
#> [379,]  2.956945e-38 1.000000e+00
#> [380,]  6.041852e-20 1.000000e+00
#> [381,]  1.000000e+00 8.452984e-23
#> [382,]  1.000000e+00 2.788317e-47
#> [383,]  2.126004e-43 1.000000e+00
#> [384,]  1.261338e-44 1.000000e+00
#> [385,]  1.000000e+00 1.712208e-42
#> [386,] 1.423486e-104 1.000000e+00
#> [387,]  3.693990e-28 1.000000e+00
#> [388,]  1.447902e-85 1.000000e+00
#> [389,]  1.868793e-32 1.000000e+00
#> [390,]  1.000000e+00 2.218183e-36
#> [391,]  2.683180e-91 1.000000e+00
#> [392,]  1.000000e+00 9.571411e-12
#> [393,]  1.000000e+00 1.129633e-22
#> [394,]  3.792266e-44 1.000000e+00
#> [395,]  5.312077e-85 1.000000e+00
#> [396,]  1.000000e+00 2.545840e-20
#> [397,] 1.121000e-137 1.000000e+00
#> [398,]  5.538811e-44 1.000000e+00
#> [399,]  1.000000e+00 1.599505e-19
#> [400,]  7.522779e-79 1.000000e+00
#> [401,]  8.540305e-49 1.000000e+00
#> [402,]  1.012917e-89 1.000000e+00
#> [403,]  1.000000e+00 5.456244e-49
#> [404,]  6.083155e-58 1.000000e+00
#> [405,]  1.000000e+00 3.111524e-38
#> [406,]  1.619648e-32 1.000000e+00
#> [407,]  5.984901e-34 1.000000e+00
#> [408,]  9.457441e-70 1.000000e+00
#> [409,]  2.608631e-29 1.000000e+00
#> [410,]  8.265988e-54 1.000000e+00
#> [411,]  1.061340e-78 1.000000e+00
#> [412,]  2.080410e-42 1.000000e+00
#> [413,]  1.000000e+00 1.267211e-28
#> [414,]  1.000000e+00 3.805574e-17
#> [415,]  1.000000e+00 9.741086e-18
#> [416,]  1.000000e+00 3.630413e-25
#> [417,] 1.943417e-157 1.000000e+00
#> [418,] 6.189249e-167 1.000000e+00
#> [419,]  2.795015e-55 1.000000e+00
#> [420,]  3.675133e-54 1.000000e+00
#> [421,]  4.722945e-24 1.000000e+00
#> [422,]  1.000000e+00 1.620596e-72
#> [423,]  6.239338e-50 1.000000e+00
#> [424,]  1.000000e+00 4.181026e-20
#> [425,]  1.495574e-96 1.000000e+00
#> [426,]  1.000000e+00 1.117070e-21
#> [427,]  4.010166e-84 1.000000e+00
#> [428,]  2.387518e-85 1.000000e+00
#> [429,]  1.000000e+00 1.895378e-52
#> [430,]  1.000000e+00 1.334282e-18
#> [431,]  1.000000e+00 1.008797e-43
#> [432,]  3.688398e-49 1.000000e+00
#> [433,]  1.000000e+00 5.384419e-14
#> [434,]  2.271740e-44 1.000000e+00
#> [435,]  1.000000e+00 5.175124e-31
#> [436,] 2.603500e-104 1.000000e+00
#> [437,]  1.199674e-57 1.000000e+00
#> [438,]  1.542363e-91 1.000000e+00
#> [439,]  8.716498e-14 1.000000e+00
#> [440,]  1.000000e+00 5.604028e-37
#> [441,]  1.000000e+00 6.573499e-53
#> [442,]  1.000000e+00 1.755039e-42
#> [443,]  1.000000e+00 8.059643e-19
#> [444,]  1.000000e+00 7.210754e-34
#> [445,]  1.000000e+00 2.359446e-47
#> [446,]  1.000000e+00 4.841903e-18
#> [447,]  1.000000e+00 1.134186e-18
#> [448,]  3.189168e-57 1.000000e+00
#> [449,]  3.737815e-67 1.000000e+00
#> [450,]  1.803892e-80 1.000000e+00
#> [451,]  6.033018e-19 1.000000e+00
#> [452,]  1.000000e+00 1.732257e-26
#> [453,]  1.000000e+00 2.315190e-22
#> [454,]  1.000000e+00 1.483887e-30
#> [455,]  3.509537e-58 1.000000e+00
#> [456,]  1.000000e+00 5.072442e-15
#> [457,]  2.582286e-63 1.000000e+00
#> [458,]  1.000000e+00 4.652250e-20
#> [459,]  1.644472e-95 1.000000e+00
#> [460,] 2.282502e-106 1.000000e+00
#> [461,]  1.000000e+00 5.124524e-36
#> [462,]  1.000000e+00 2.192286e-15
#> [463,]  1.000000e+00 3.832503e-35
#> [464,]  1.000000e+00 9.685785e-70
#> [465,]  1.216099e-42 1.000000e+00
#> [466,]  1.000000e+00 1.791986e-28
#> [467,]  1.000000e+00 1.164112e-22
#> [468,]  1.035617e-77 1.000000e+00
#> [469,]  6.676789e-38 1.000000e+00
#> [470,]  4.640977e-72 1.000000e+00
#> [471,]  3.100709e-60 1.000000e+00
#> [472,]  3.911329e-21 1.000000e+00
#> [473,]  1.217322e-49 1.000000e+00
#> [474,]  1.000000e+00 6.356748e-14
#> [475,]  1.182057e-22 1.000000e+00
#> [476,]  1.000000e+00 1.360857e-70
#> [477,]  1.000000e+00 1.050675e-23
#> [478,]  1.096360e-46 1.000000e+00
#> [479,]  2.512030e-22 1.000000e+00
#> [480,]  1.000000e+00 2.704619e-25
#> [481,]  1.586166e-94 1.000000e+00
#> [482,]  7.320179e-21 1.000000e+00
#> [483,] 3.851125e-102 1.000000e+00
#> [484,]  1.000000e+00 7.860236e-26
#> [485,]  1.000000e+00 1.504904e-31
#> [486,]  3.639516e-71 1.000000e+00
#> [487,]  1.000000e+00 1.207697e-60
#> [488,]  1.000000e+00 9.024110e-55
#> [489,]  1.972417e-73 1.000000e+00
#> [490,] 1.398032e-155 1.000000e+00
#> [491,]  1.000000e+00 3.665951e-30
#> [492,]  4.753319e-17 1.000000e+00
#> [493,]  1.009396e-99 1.000000e+00
#> [494,]  1.000000e+00 1.537845e-23
#> [495,]  1.000000e+00 1.536538e-58
#> [496,]  2.123155e-28 1.000000e+00
#> [497,]  1.000000e+00 3.435269e-43
#> [498,]  3.331525e-96 1.000000e+00
#> [499,]  4.892128e-40 1.000000e+00
#> [500,]  1.000000e+00 9.960794e-47