Simulate from the cellassign model

Simulate RNA-seq counts from the cell-assign model

simulate_cellassign(rho, s, pi, delta, B = 20, a, beta, X = NULL,
  min_Y = 0, max_Y = 1000)

Arguments

rho
A gene by cell type binary matrix relating markers to cell types

s
A vector of cell-specific size factors

pi
An ordinal vector relating each cell to its true marker type assignment

delta
Gene by cell type matrix delta (all entries with corresponding zeros in rho will be ignored)

B
Granularity of spline-based fitting of dispersions

a
Alpha parameters for spline inference of dispersions

beta
A gene by covariate vector of coefficients - the first column should correspond to the intercept (baseline expression) values

X
A cell by covariate matrix of covariates - the intercept column will always be added.

b
Beta parameters for spline inference of dispersions

Value

An N by G matrix of simulated counts

Details

The number of genes, cells, and cell types is automatically inferred from the dimensions of rho (gene by cell-type) and s (vector of length number of cells). The specification of X is optional - a column of ones will always be added as an intercept.